Add tst mocks

+++ src/model.rs
@@ -2,7 +2,7 @@ use chrono::{DateTime, Utc};
 
 pub struct Session {
     pub id: String,
-    pub user_id: i64,
+    pub user_id: i32,
     pub access_token: String,
     pub refresh_token: String,
 }
